+++ src/cqspoolterrain/pipe/__init__.py
@@ -2,3 +2,5 @@ from .pipeFace import pipe_face
 from .straight import straight
 from .curve import curve
 from .end import end
+from .connector import connector
+from .hatch import hatch